pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "replicate-runner"
version = "0.1.0"
description = "CLI tool for running Replicate AI models and publishing LoRAs to HuggingFace"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "replicate>=0.25.0",
    "huggingface-hub>=0.20.0",
]

[project.scripts]
replicate-runner = "replicate_runner.main:run"

[tool.uv]
dev-dependencies = [
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
